add query selector, complete database generator, working on optimizing querying functions
