+++ .gitmodules
@@ -1,3 +1,3 @@
 [submodule "tools/cpptqdm"]
 	path = tools/cpptqdm
-	url = https://github.com/aminnj/cpptqdm.git
+	url = https://github.com/yaoyuan0553/cpptqdm.git
