--- tools/Database/src/DatabaseQuery.cpp
@@ -1,5 +1,0 @@
-//
-// Created by yuan on 8/17/19.
-//
-
-#include "DatabaseQuery.h"
